requirements.txt
python-dotenv
httpx
asyncio
requests